requirements.txt
python-telegram-bot==21.7
APScheduler==3.10.4
SQLAlchemy==2.0.35
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
